created edit profile component and created base sign up form
