+++ package.json
@@ -21,6 +21,7 @@
     "react": "^16.13.1",
     "react-apollo": "^3.1.5",
     "react-dom": "^16.13.1",
+    "react-hook-form": "^6.0.8",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
     "sweetalert": "^2.1.2"
